+++ src/Kit/Text/IString.h
@@ -55,32 +55,32 @@ public:
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( int num ) noexcept = 0;
+    virtual IString& operator=( int16_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( unsigned int num ) noexcept = 0;
+    virtual IString& operator=( uint16_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( long num ) noexcept = 0;
+    virtual IString& operator=( int32_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( long long num ) noexcept = 0;
+    virtual IString& operator=( uint32_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( unsigned long num ) noexcept = 0;
+    virtual IString& operator=( int64_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and stores its in
         the internal storage of the object.
      */
-    virtual IString& operator=( unsigned long long num ) noexcept = 0;
+    virtual IString& operator=( uint64_t num ) noexcept = 0;
 
     /** This method copies the content of the source string,
         up to 'n' characters, into internal storage of the object.
@@ -123,32 +123,32 @@ public:
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( int num ) noexcept = 0;
+    virtual IString& operator+=( int16_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( unsigned int num ) noexcept = 0;
+    virtual IString& operator+=( uint16_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( long num ) noexcept = 0;
+    virtual IString& operator+=( int32_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( long long num ) noexcept = 0;
+    virtual IString& operator+=( uint32_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( unsigned long num ) noexcept = 0;
+    virtual IString& operator+=( int64_t num ) noexcept = 0;
 
     /** Converts the 'number' to a string and appends its to the
         end of the string.
      */
-    virtual IString& operator+=( unsigned long long num ) noexcept = 0;
+    virtual IString& operator+=( uint64_t num ) noexcept = 0;
 
     /** This method appends the content of the source string,
         up to 'n' characters, into internal storage of the object.
